reference-implementation/requirements.txt
fastapi
uvicorn
pyyaml
jsonschema
pytest
httpx

